+++ src/lib/kotlin/slatekit-data/src/main/kotlin/slatekit/data/encoders/DoubleEncoder.kt
@@ -1,8 +1,8 @@
-package slatekit.orm.databases.converters
+package slatekit.data.encoders
 
-import slatekit.orm.core.SqlEncoder
+
 import slatekit.common.Record
-import slatekit.orm.Consts
+import slatekit.data.Consts
 
 class DoubleEncoder : SqlEncoder<Double> {
